Move each action's specs to their own file
