+++ spec/flipper/ui/actions/add_feature_spec.rb
@@ -1,0 +1,27 @@
+require 'helper'
+require 'rack/test'
+require 'flipper'
+require 'flipper/adapters/memory'
+require 'flipper/ui/actor'
+
+describe Flipper::UI::Actions::AddFeature do
+  include Rack::Test::Methods
+
+  let(:adapter) { Flipper::Adapters::Memory.new }
+  let(:flipper) { Flipper.new(adapter) }
+  let(:app)     { Flipper::UI.app(flipper) }
+
+  describe "GET /features/new" do
+    before do
+      get "/features/new"
+    end
+
+    it "responds with success" do
+      last_response.status.should be(200)
+    end
+
+    it "renders template" do
+      last_response.body.should include('<form action="/features" method="post">')
+    end
+  end
+end
